Fix dependencies 2/2: restructure everything Signed-off-by: Brendan Cronin <brendan.cronin@mercedes-benz.com>
