+++ edc-extensions/transferprocess-sftp-common/pom.xml
@@ -18,7 +18,7 @@
     <parent>
         <groupId>org.eclipse.tractusx.edc.extensions</groupId>
         <artifactId>edc-extensions</artifactId>
-        <version>0.1.4-SNAPSHOT</version>
+        <version>0.1.5-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
