+++ edc-extensions/transferprocess-sftp-common/src/main/java/net/catenax/edc/trasnferprocess/sftp/common/SftpLocation.java
@@ -12,21 +12,19 @@
  *
  */
 
-package net.catenax.edc.transferprocess.sftp.provisioner;
+package net.catenax.edc.trasnferprocess.sftp.common;
 
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 
-import java.security.KeyPair;
-
 @Builder
 @Getter
-@ToString(of = "name")
+@ToString
 @EqualsAndHashCode
-public class SftpUser {
-    private final String name;
-    private final String password;
-    private final KeyPair keyPair;
+public class SftpLocation {
+    private final String host;
+    private final Integer port;
+    private final String path;
 }
