+++ edc-extensions/transferprocess-sftp-common/src/main/java/net/catenax/edc/trasnferprocess/sftp/common/SftpUserFactory.java
@@ -12,7 +12,7 @@
  *
  */
 
-package net.catenax.edc.transferprocess.sftp.provisioner;
+package net.catenax.edc.trasnferprocess.sftp.common;
 
 public interface SftpUserFactory {
     SftpUser createSftpUser(String transferProcessId);
